dihedral updates to functions.py - DA calculates dihedral angle (or torsional angle?) using 4 points - V_D calculates dihedral potential
V_D is split into two different functional forms because I don't know
which one we should use. There are also many blank spots in the
parameters as I am unsure how we are finding these as of yet.

In the DA function I assumed we are sending our positions array and
a reference atom number as input. This structure needs work to match our
dictionaries we want to use. I am unsure of how we are implementing this
yet so this function only serves as the calculation concept.